# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Nexora API
  version: 1.0.0
  description: 'Nexora is a multi-currency digital asset platform built on Stellar. It provides

    wallets, instant transfers, FX conversions, scheduled recurring payments,

    batch disbursements, fiat on/off-ramps, and a developer-friendly webhook system.


    All endpoints under `/v1` require a Bearer API key in the `Authorization` header.


    ## API key scopes


    An API key can be limited to specific resources and actions. Scopes have the

    form `<resource>:<action>`, where action is `read` or `write`:


    | Resource | Scopes | Covers |

    |---|---|---|

    | wallets | `wallets:read`, `wallets:write` | `/v1/wallets` (create, trustlines, contract wallet changes) |

    | transfers | `transfers:read`, `transfers:write` | `/v1/transfers`, `/v1/transactions`, `/v1/schedules`, `/v1/refunds` |

    | batches | `batches:read`, `batches:write` | `/v1/transfers/batch` |

    | webhooks | `webhooks:read`, `webhooks:write` | `/v1/webhooks` |

    | reports | `reports:read`, `reports:write` | `/v1/usage` (`reports:write` is reserved for report generation) |


    Other scopes: `keys:*`, `audit:read`, `fiat:*`, `fx:*`, `fees:read`,

    `beneficiaries:*`, `compliance:*`. `<resource>:*` grants both actions on one

    resource and `*` grants everything.


    On scoped resource groups, `GET`/`HEAD` need the `read` scope and every other

...

...

...

    path only; never the key itself).


...

...

...

...

...

...

    '
  contact:
...
servers:
...
security:
- BearerAuth: []
paths:
...
  /v1/schedules:
    post:
      operationId: createSchedule
      summary: Create a recurring tenant payout schedule
      description: Creates a tenant-scoped recurring payout that runs daily, weekly, or monthly. Requires transfers:write scope and a UUID v4 Idempotency-Key; retries with the same key and body replay the original response.
      tags:
      - Schedules
      parameters:
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          format: uuid
        description: UUID v4 key for this schedule creation; reuse it only when retrying the same request.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateScheduleRequest'
            example:
              from_wallet_id: 3f1c2b8e-9a4d-4d3f-8f4e-2b6c1a9d0e77
              to_wallet_id: 4f1c2b8e-9a4d-4d3f-8f4e-2b6c1a9d0e77
              asset: USDC
              amount: '100.0000000'
              frequency: monthly
              timezone: Africa/Lagos
              missed_run_policy: skip
              start_date: '2026-02-01T00:00:00Z'
              end_date: '2026-12-31T23:59:59Z'
      responses:
        '201':
          description: Schedule created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
    get:
      operationId: listSchedules
      summary: List schedules
      description: Returns up to 100 recurring payout schedules belonging to the authenticated tenant. Requires transfers:read scope.
      tags:
      - Schedules
      responses:
        '200':
          description: List of schedules
          content:
            application/json:
              schema:
                type: object
                properties:
                  schedules:
                    type: array
                    items:
                      $ref: '#/components/schemas/ScheduleResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/schedules/{id}/runs:
    get:
      operationId: listScheduleRuns
      summary: List payout attempts for a schedule
      description: Returns the authenticated tenant's execution history, newest occurrence first. Requires transfers:read scope. Limit defaults to 20 and is capped at 100; offset defaults to 0.
      tags:
      - Schedules
      parameters:
      - $ref: '#/components/parameters/ScheduleID'
      - name: limit
        in: query
        schema:
          type: integer
          minimum: 1
          maximum: 100
      - name: offset
        in: query
        schema:
          type: integer
          minimum: 0
      responses:
        '200':
          description: Payout run history
          content:
            application/json:
              schema:
                type: object
                properties:
                  runs:
                    type: array
                    items:
                      $ref: '#/components/schemas/ScheduleRunResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/schedules/{id}:
    patch:
      operationId: updateSchedule
      summary: Update a schedule
      description: Updates the amount, recurrence settings, or status. Requires transfers:write scope.
      tags:
      - Schedules
      parameters:
      - $ref: '#/components/parameters/ScheduleID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateScheduleRequest'
            example:
              status: paused
              amount: '150.0000000'
      responses:
        '200':
          description: Schedule updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      operationId: cancelSchedule
      summary: Cancel a schedule
      description: Cancels a recurring tenant payout schedule. Requires transfers:write scope. Cannot be undone.
      tags:
      - Schedules
      parameters:
      - $ref: '#/components/parameters/ScheduleID'
      responses:
        '204':
          description: Schedule cancelled
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
      description: 'Pass a tenant API key (`sk_live_...` or `sk_test_...`) or JWT bearer token.

        In Swagger UI, enter your token in the Authorize dialog (the ''Bearer '' prefix is added automatically).

        '
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: 'Pass an API key as `Bearer <key>` in the `Authorization` header.

        '
    AdminAuth:
      type: http
      scheme: bearer
      bearerFormat: admin token
      description: Administrative bearer token.
  parameters:
...
    ScheduleID:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Schedule ID
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: BAD_REQUEST
              message: invalid request body
              status: 400
              request_id: req-example
    Unauthorized:
      description: Unauthorized — missing or invalid API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: UNAUTHORIZED
              message: missing or invalid authorization header
              status: 401
              request_id: req-example
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: NOT_FOUND
              message: resource not found
              status: 404
              request_id: req-example
    Forbidden:
      description: Forbidden — admin access required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: FORBIDDEN
              message: admin access required
              status: 403
              request_id: req-example
    Conflict:
      description: Conflict — a request with this Idempotency-Key is still in flight
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: REQUEST_IN_PROGRESS
              message: a request with this Idempotency-Key is still in progress
              status: 409
              request_id: req-example
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: INTERNAL_ERROR
              message: an unexpected error occurred
              status: 500
              request_id: req-example
  schemas:
...
    ErrorResponse:
      type: object
      required:
      - error
      properties:
        error:
          type: object
          required:
          - code
          - message
          - status
          - request_id
          properties:
            code:
              type: string
              enum:
              - BAD_REQUEST
              - ANCHOR_DEPOSIT_FAILED
              - ANCHOR_REGISTRATION_FAILED
              - ANCHOR_TRANSACTION_NOT_FOUND
              - ANCHOR_WITHDRAWAL_FAILED
              - API_KEY_NOT_FOUND
              - API_KEY_REVOKED
              - API_KEY_EXPIRED
              - BENEFICIARY_EXISTS
              - BENEFICIARY_NOT_ACTIVE
              - BENEFICIARY_NOT_ALLOWED
              - CONFIG_NOT_FOUND
              - CONFLICT
              - COOLING_OFF
              - CROSS_MODE_KEY_OPERATION
              - ENVIRONMENT_REQUIRED
              - IDEMPOTENCY_KEY_REQUIRED
              - IDEMPOTENCY_RECORD_PERSISTENCE_FAILED
              - IDEMPOTENCY_RESPONSE_TOO_LARGE
              - INSUFFICIENT_SCOPE
              - INVALID_ENVIRONMENT_MODE
              - INVALID_PREDICATE
              - INVALID_REQUEST
              - LAST_ORG_OWNER
              - METHOD_NOT_ALLOWED
              - MISSING_TRUSTLINE
              - NOT_IMPLEMENTED
              - PAYMENT_LINK_UNAVAILABLE
              - PREDICATE_NOT_SATISFIABLE
              - RATE_LIMITED
              - REFUND_CONFLICT
              - TEST_MODE_REQUIRED
              - TRANSFER_ID_INVALID
              - TRANSFER_ID_REQUIRED
              - TRANSFER_NOT_CANCELLABLE
              - UNSUPPORTED_FIAT_CURRENCY
              - WALLET_ID_REQUIRED
              - NOT_FOUND
              - UNAUTHORIZED
              - FORBIDDEN
              - INSUFFICIENT_SCOPE
              - INTERNAL_ERROR
              - INSUFFICIENT_SWEEPABLE_BALANCE
              - QUOTE_EXPIRED
              - QUOTE_ALREADY_USED
              - REVIEW_ALREADY_DECIDED
              - CLAIMABLE_BALANCE_NOT_PENDING
              - REQUEST_IN_PROGRESS
              description: Machine-readable error code
            message:
              type: string
              description: Human-readable error message
            status:
              type: integer
              description: HTTP status code; matches the response status
              minimum: 400
              maximum: 599
            request_id:
              type: string
              description: Effective X-Request-ID response header value
...
    CreateScheduleRequest:
      type: object
      required:
      - from_wallet_id
      - to_wallet_id
      - asset
      - amount
      - frequency
      - start_date
      properties:
        from_wallet_id:
          type: string
          format: uuid
        to_wallet_id:
          type: string
          format: uuid
        asset:
          type: string
        amount:
          type: string
          description: Positive decimal amount
        frequency:
          type: string
          enum:
          - daily
          - weekly
          - monthly
        timezone:
          type: string
          description: IANA time zone; defaults to UTC
        missed_run_policy:
          type: string
          enum:
          - skip
          - run_once
          description: Defaults to run_once. skip advances past elapsed occurrences; run_once makes one catch-up payout.
        start_date:
          type: string
          format: date-time
          description: RFC3339 timestamp
        end_date:
          type: string
          format: date-time
          description: Optional RFC3339 timestamp; must not precede start_date
    ScheduleRunResponse:
      type: object
      properties:
        id:
          type: string
        schedule_id:
          type: string
        expected_run_at:
          type: string
          format: date-time
        status:
          type: string
          enum:
          - pending
          - running
          - succeeded
          - failed
          - skipped
          - cancelled
        transaction_id:
          type: string
          nullable: true
        error:
          type: string
          description: Safe explanation shown only when the run failed
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
    UpdateScheduleRequest:
      type: object
      properties:
        status:
          type: string
          enum:
          - active
          - paused
          description: Pause or resume the schedule
        amount:
          type: string
          description: New amount as decimal string
        frequency:
          type: string
          enum:
          - daily
          - weekly
          - monthly
        end_date:
          type: string
          format: date-time
        timezone:
          type: string
        missed_run_policy:
          type: string
          enum:
          - skip
          - run_once
    ScheduleResponse:
      type: object
      properties:
        id:
          type: string
        from_wallet_id:
          type: string
        to_wallet_id:
          type: string
        asset:
          type: string
        amount:
          type: string
        frequency:
          type: string
          enum:
          - daily
          - weekly
          - monthly
        timezone:
          type: string
        missed_run_policy:
          type: string
          enum:
          - skip
          - run_once
        next_run_at:
          type: string
          format: date-time
        end_at:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
          enum:
          - active
          - processing
          - failed
          - paused
          - cancelled
          - completed
        created_at:
          type: string
          format: date-time
...